Completes #21 and adds abstractions to ExpyTableView class
