+++ Example/ExpyTableView/MainViewController.swift
@@ -9,9 +9,8 @@
 import UIKit
 
 class MainViewController: UIViewController {
-
-	override func viewWillAppear(_ animated: Bool) {
-		super.viewWillAppear(animated)
+	override func viewDidLoad() {
+		super.viewDidLoad()
 		navigationItem.title = "Expy Examples"
 	}
 }
